service/requirements.txt
Django==3.1.2
djangorestframework==3.11.0
requests==2.27.1